room making and updating works but need to test more
